Rename `services` to `models`
